Removed closure library. Stopped using advanced optimizations with closure compiler.
Advanced optimizations won't work without externs and work at all with angular so I had to ditch it. I also discovered that the closure compiler can deal with goog.require and goog.provide without the closure library which was all I wanted from it anyway, so I removed the closure library.

There is still a problem with how the compile build rule is set up. I want to experiment with using goog.provide and goog.require for third party libs.